--- tests/Concept2/Exception/__init__.py
@@ -1,3 +1,0 @@
-"""
-tests.PyRow.Concept2.Exception
-"""
