--- tests/Concept2/__init__.py
@@ -1,3 +1,0 @@
-"""
-tests.PyRow.Concept2
-"""
