+++ tests/csafe_cmd.py
